Separate connection configuration logic into mixin
Connection configuration logic has to be reused in Provider
therefore it's best to put it into mixin.
